+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 3.15.9 (September 19, 2022)
+
+### Fixed
+
+- Fix WAI-ARIA rules for Tooltip component
+- Fix slide animation in Sortable component
+- Fix input fields reverting to previous value while sorting in Sortable component
+
 ## 3.15.8 (September 14, 2022)
 
 ### Fixed